Moving concurrent-search out of sandbox plugin to core behind feature flag (#7203)
