--- sandbox/plugins/concurrent-search/src/main/java/org/opensearch/search/package-info.java
@@ -1,12 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-/**
- * The implementation of the experimental plugin which implements the concurrent search over Apache Lucene segments.
- */
-package org.opensearch.search;
