--- sandbox/plugins/concurrent-search/src/main/java/org/opensearch/search/query/package-info.java
@@ -1,12 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-/**
- * {@link org.opensearch.search.query.QueryPhaseSearcher} implementation for concurrent search
- */
-package org.opensearch.search.query;
