+++ server/src/main/java/org/opensearch/common/settings/FeatureFlagSettings.java
@@ -38,7 +38,8 @@ public class FeatureFlagSettings extends AbstractScopedSettings {
                 FeatureFlags.REMOTE_STORE_SETTING,
                 FeatureFlags.EXTENSIONS_SETTING,
                 FeatureFlags.IDENTITY_SETTING,
-                FeatureFlags.SEARCH_PIPELINE_SETTING
+                FeatureFlags.SEARCH_PIPELINE_SETTING,
+                FeatureFlags.CONCURRENT_SEGMENT_SEARCH_SETTING
             )
         )
     );
