+++ server/src/main/java/org/opensearch/search/SearchModule.java
@@ -35,7 +35,6 @@ package org.opensearch.search;
 import org.apache.lucene.search.BooleanQuery;
 import org.opensearch.common.NamedRegistry;
 import org.opensearch.common.Nullable;
-import org.opensearch.core.ParseField;
 import org.opensearch.common.geo.GeoShapeType;
 import org.opensearch.common.geo.ShapesAvailability;
 import org.opensearch.common.io.stream.NamedWriteableRegistry;
@@ -43,8 +42,10 @@ import org.opensearch.common.io.stream.NamedWriteableRegistry.Entry;
 import org.opensearch.common.io.stream.Writeable;
 import org.opensearch.common.settings.Setting;
 import org.opensearch.common.settings.Settings;
-import org.opensearch.core.xcontent.NamedXContentRegistry;
+import org.opensearch.common.util.FeatureFlags;
 import org.opensearch.common.xcontent.ParseFieldRegistry;
+import org.opensearch.core.ParseField;
+import org.opensearch.core.xcontent.NamedXContentRegistry;
 import org.opensearch.core.xcontent.XContentParser;
 import org.opensearch.index.query.BoolQueryBuilder;
 import org.opensearch.index.query.BoostingQueryBuilder;
@@ -255,6 +256,7 @@ import org.opensearch.search.fetch.subphase.highlight.HighlightPhase;
 import org.opensearch.search.fetch.subphase.highlight.Highlighter;
 import org.opensearch.search.fetch.subphase.highlight.PlainHighlighter;
 import org.opensearch.search.fetch.subphase.highlight.UnifiedHighlighter;
+import org.opensearch.search.query.ConcurrentQueryPhaseSearcher;
 import org.opensearch.search.query.QueryPhase;
 import org.opensearch.search.query.QueryPhaseSearcher;
 import org.opensearch.search.rescore.QueryRescorerBuilder;
@@ -292,6 +294,7 @@ import static java.util.Collections.unmodifiableList;
 import static java.util.Collections.unmodifiableMap;
 import static java.util.Objects.requireNonNull;
 import static org.opensearch.index.query.CommonTermsQueryBuilder.COMMON_TERMS_QUERY_DEPRECATION_MSG;
+import static org.opensearch.threadpool.ThreadPool.Names.INDEX_SEARCHER;
 
 /**
  * Sets up things that can be done at search time like queries, aggregations, and suggesters.
@@ -1252,6 +1255,9 @@ public class SearchModule {
             }
         }
 
+        if (searcher == null && FeatureFlags.isEnabled(FeatureFlags.CONCURRENT_SEGMENT_SEARCH)) {
+            searcher = new ConcurrentQueryPhaseSearcher();
+        }
         return searcher;
     }
 
@@ -1270,6 +1276,9 @@ public class SearchModule {
             }
         }
 
+        if (provider == null && FeatureFlags.isEnabled(FeatureFlags.CONCURRENT_SEGMENT_SEARCH)) {
+            provider = (ThreadPool threadPool) -> threadPool.executor(INDEX_SEARCHER);
+        }
         return provider;
     }
 
@@ -1282,6 +1291,6 @@ public class SearchModule {
     }
 
     public @Nullable ExecutorService getIndexSearcherExecutor(ThreadPool pool) {
-        return (indexSearcherExecutorProvider == null) ? null : indexSearcherExecutorProvider.getExecutor(pool);
+        return (indexSearcherExecutorProvider != null) ? indexSearcherExecutorProvider.getExecutor(pool) : null;
     }
 }
